Separate webpack config into separate files
