+++ .eslintignore
@@ -2,6 +2,7 @@
 node_modules
 build
 webpack.config.js
+config/*
 server.js
 postcss.config.js
 scripts/*
