+++ config/webpack.heroku.config.js
@@ -1,0 +1,4 @@
+const baseConfig = require("./webpack.common.config");
+const merge = require("webpack-merge");
+
+module.exports = merge(baseConfig, {});
